libvdsk: Change disk specific auxiliary data referencing
Up until now we used a void pointer to the disk specific data in the
vdsk structure.

Now we allocate a bigger area and get to the disk specific data by
subtracting the size of the structure from the pointer to the vdsk
structure.

In order to do this, we need to store the size of the structure in the
vdsk_format structure.

This decreases the amount of indirections we have to do for each
operation.

Signed-off-by: Sergiu Weisz <sergiu121@gmail.com>